backend/revectorize-documents/requirements.txt
psycopg2-binary==2.9.9
PyJWT==2.8.0
openai==1.12.0
requests==2.31.0